GX2-4687 | rename POM and Fixture
