+++ cypress/support/pages/GX2-4787-stickerPage.js
